Implementation of mobile guide menu
