+++ apps/nextjs-website/src/components/atoms/GuideMenu/GuideMenu.tsx
@@ -1,239 +1,117 @@
 'use client';
-import React, { ReactNode } from 'react';
-import NextLink from 'next/link';
+import React, { useCallback, useState } from 'react';
 import Typography from '@mui/material/Typography';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import ExpandLessIcon from '@mui/icons-material/ExpandLess';
-import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import { usePathname } from 'next/navigation';
-import { TreeItem, TreeView, treeItemClasses } from '@mui/lab';
-import { styled } from '@mui/material/styles';
-import { parseMenu } from 'gitbook-docs/parseMenu';
-import { RenderingComponents, renderMenu } from 'gitbook-docs/renderMenu';
-import Dropdown from '@/components/atoms/Dropdown/Dropdown';
-import { ExpandLess, ExpandMore } from '@mui/icons-material';
-import { Box, useTheme } from '@mui/material';
-import { translations } from '@/_contents/translations';
-import { useScrollUp } from '../ProductHeader/useScrollUp';
+import ArrowDropDownOutlinedIcon from '@mui/icons-material/ArrowDropDownOutlined';
+import {
+  Collapse,
+  IconButton,
+  Stack,
+  Theme,
+  useMediaQuery,
+  useTheme,
+} from '@mui/material';
 import { SITE_HEADER_HEIGHT } from '@/components/molecules/SiteHeader/SiteHeader';
+import { useScrollUp } from '../ProductHeader/useScrollUp';
+import GuideMenuItems, { type GuideMenuItemsProps } from './Menu';
 
-type GuideMenuProps = {
-  linkPrefix: string;
-  assetsPrefix: string;
-  menu: string;
+type GuideMenuProps = GuideMenuItemsProps & {
   guideName: string;
-  versionName: string;
-  versions: { name: string; path: string }[];
 };
 
-export const PRODUCT_HEADER_HEIGHT = 80;
-
-const StyledTreeItem = styled(TreeItem)(({ theme }) => ({
-  [`&`]: {
-    '--x': 32,
-  },
-  [`& .${treeItemClasses.content}`]: {
-    boxSizing: 'border-box',
-    flexDirection: 'row-reverse',
-    width: '100%',
-    paddingTop: 0,
-    paddingBottom: 0,
-    paddingLeft: 0,
-    paddingRight: 32,
-    alignItems: 'space-between',
-  },
-  [`& .${treeItemClasses.content}:has(.${treeItemClasses.iconContainer}:empty)`]:
-    {
-      paddingRight: 0,
-    },
-  [`& .${treeItemClasses.iconContainer}`]: {
-    justifyContent: 'flex-end',
-    marginRight: 0,
-    paddingRight: 0,
-    paddingLeft: 0,
-  },
-  [`& .${treeItemClasses.iconContainer}:empty`]: {
-    display: 'none',
-  },
-  [`& .${treeItemClasses.content} > .${treeItemClasses.label}`]: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'stretch',
-    position: 'relative',
-  },
-  [`& .${treeItemClasses.content} > .${treeItemClasses.label} > a`]: {
-    paddingTop: 16,
-    paddingBottom: 16,
-    paddingRight: 32,
-  },
-  [`& ul`]: {
-    paddingLeft: 0,
-    '--y': 'calc(var(--x) + 0)',
-  },
-  [`& li`]: {
-    '--x': 'calc(var(--y) + 24)',
-  },
-  ['& a']: {
-    paddingLeft: 'calc(1px * var(--x))',
-  },
-  [`& .${treeItemClasses.group}`]: {
-    marginLeft: 0,
-    marginRight: 0,
-  },
-  [`& .${treeItemClasses.group} .${treeItemClasses.label}`]: {
-    paddingLeft: 0,
-    paddingRight: 0,
-  },
-  [`& .${treeItemClasses.label}`]: {
-    padding: 0,
-    paddingLeft: 0,
-  },
-  [`& .${treeItemClasses.root}`]: {
-    margin: 0,
-    paddingLeft: 0,
-  },
-  [`& .${treeItemClasses.selected}`]: {
-    borderRight: `2px solid ${theme.palette.primary.dark}`,
-  },
-  [`& .${treeItemClasses.selected} > .${treeItemClasses.label} > *`]: {
-    color: theme.palette.primary.dark,
-  },
-}));
-
-const components: RenderingComponents<ReactNode> = {
-  Item: ({ href, title, children }) => {
-    const label = (
-      <Typography
-        variant='sidenav'
-        component={NextLink}
-        href={href}
-        style={{ textDecoration: 'none' }}
-      >
-        {title}
-      </Typography>
-    );
-
-    return (
-      <StyledTreeItem
-        key={href}
-        nodeId={href}
-        label={label}
-        disabled={false}
-        icon={href.startsWith('http') ? <OpenInNewIcon /> : undefined}
-      >
-        {children}
-      </StyledTreeItem>
-    );
-  },
-  Title: ({ children }) => (
-    <Typography
-      color='text.secondary'
-      style={{
-        paddingLeft: 32,
-        paddingTop: 24,
-        paddingBottom: 0,
-        textDecoration: 'none',
-        fontSize: 14,
-        fontWeight: 700,
-      }}
-      textTransform='uppercase'
-    >
-      {children}
-    </Typography>
-  ),
+export const PRODUCT_HEADER_HEIGHT = {
+  xs: 60,
+  md: 90,
+  lg: 75,
 };
 
-const GuideMenu = ({
-  menu,
-  assetsPrefix,
-  linkPrefix,
-  guideName,
-  versionName,
-  versions,
-}: GuideMenuProps) => {
+const GuideMenu = ({ guideName, ...menuProps }: GuideMenuProps) => {
+  const [open, setOpen] = useState(false);
+  const isDesktop = useMediaQuery((theme: Theme) => theme.breakpoints.up('lg'));
   const { palette } = useTheme();
-  const { shared } = translations;
   const scrollUp = useScrollUp();
   const currentPath = usePathname();
   const segments = currentPath.split('/');
   const expanded = segments.map((_, i) => segments.slice(0, i + 1).join('/'));
-  const top = scrollUp
-    ? SITE_HEADER_HEIGHT + PRODUCT_HEADER_HEIGHT
-    : PRODUCT_HEADER_HEIGHT;
+  // const top = scrollUp
+  //   ? SITE_HEADER_HEIGHT + PRODUCT_HEADER_HEIGHT
+  //   : PRODUCT_HEADER_HEIGHT;
+
+  const getTopValue = (breakpoint: 'xs' | 'md' | 'lg') => {
+    const top = scrollUp
+      ? SITE_HEADER_HEIGHT + PRODUCT_HEADER_HEIGHT[breakpoint]
+      : PRODUCT_HEADER_HEIGHT[breakpoint];
+    return top;
+  };
+
+  const height = `calc(100vh - ${getTopValue('lg')}px)`;
+
+  const handleClick = useCallback(() => {
+    setOpen((prev) => !prev);
+  }, []);
+
+  const items = (
+    <GuideMenuItems
+      {...menuProps}
+      currentPath={currentPath}
+      expanded={expanded}
+    />
+  );
+
   return (
-    <Box
+    <Stack
       sx={{
         backgroundColor: palette.grey[50],
-        display: 'flex',
-        flexDirection: 'column',
         flexShrink: 0,
-        position: { lg: 'sticky' },
-        top: { lg: top },
-        height: { lg: `calc(100vh - ${top}px)` },
+        position: 'sticky',
+        top: {
+          xs: getTopValue('xs'),
+          md: getTopValue('md'),
+          lg: getTopValue('lg'),
+        },
+        height: { xs: 'auto', lg: height },
+        maxHeight: { xs: 350, lg: height },
         overflowY: 'auto',
         transition: 'all 0.5s linear',
         scrollbarWidth: 'thin',
         width: { lg: '347px' },
       }}
     >
-      <Box
+      <Stack
         sx={{
-          display: 'flex',
-          flexDirection: 'column',
-          padding: '80px 0',
+          padding: { lg: '80px 0' },
           flexGrow: { lg: 0 },
           flexShrink: { lg: 0 },
         }}
       >
-        <Typography
-          variant='h6'
-          sx={{
-            padding: '16px 32px',
-            verticalAlign: 'middle',
-          }}
+        <Stack
+          direction='row'
+          alignItems='center'
+          justifyContent='flex-start'
+          onClick={isDesktop ? undefined : handleClick}
+          sx={{ padding: '16px 24px', cursor: 'pointer' }}
         >
-          {guideName}
-        </Typography>
-        <Dropdown
-          label={`${shared.version} ${versionName}`}
-          items={versions.map((version) => ({
-            href: version.path,
-            label: version.name,
-          }))}
-          icons={{ opened: <ExpandLess />, closed: <ExpandMore /> }}
-          buttonStyle={{
-            color: palette.action.active,
-            display: 'flex',
-            justifyContent: 'space-between',
-            padding: '16px 32px',
-          }}
-          menuStyle={{
-            style: {
-              width: '347px',
-              maxWidth: '347px',
-              left: 0,
-              right: 0,
-            },
-          }}
-          menuAnchorOrigin={{
-            vertical: 'bottom',
-            horizontal: 'center',
-          }}
-        />
-        <TreeView
-          defaultCollapseIcon={<ExpandLessIcon />}
-          defaultExpanded={expanded}
-          selected={currentPath}
-          defaultExpandIcon={<ExpandMoreIcon />}
-        >
-          {renderMenu(
-            parseMenu(menu, { assetsPrefix, linkPrefix }),
-            React,
-            components
-          )}
-        </TreeView>
-      </Box>
-    </Box>
+          <Typography
+            variant='h6'
+            sx={{
+              verticalAlign: 'middle',
+            }}
+          >
+            {guideName}
+          </Typography>
+          <IconButton size='small' sx={{ display: { lg: 'none' } }}>
+            <ArrowDropDownOutlinedIcon sx={{ width: 24, height: 24 }} />
+          </IconButton>
+        </Stack>
+        {isDesktop ? (
+          items
+        ) : (
+          <Collapse appear={open} in={open}>
+            {items}
+          </Collapse>
+        )}
+      </Stack>
+    </Stack>
   );
 };
